Change `usedCapturedValues` in `BailOutInfo` to a pointer for consistency
As part of `LazyBailOut`, we want to perform a *shallow* copy of a `BailOutInfo` which includes sharing all field pointers. However, `usedCapturedValues` is the only field in `BailOutInfo` that is not a pointer and is initialized together with `CapturedValues`, which makes it more complicated to do so as we have to perform manual copy of each field inside `CapturedValues`. With this change, we are more consistent with what we already have and also make it easier for future `LazyBailOut` work.