Update params.js test to not kill the process during execution
The call to WScript.Quit caused the process to terminate while script was executing, which now breaks some of the cleanup work added with this set of changes.
The fix is to make an else block for the work that would have been short-circuited, which allows the stack to unwind safely.