Fixes asserts in Speedometer - Some asserts in inline caches confused Null and Missing. CEWOs may produce the latter. - Missing property results should not be cached by inline caches if comes from CEWO when implicit calls are disabled - that may change without type changes. - rationalizing the marshaling story in CEWO. - The idea is to _not_ marshal on Get and the like, but on Set/Define marshal to the script contect of the containing CEWO.
The latter scenario arises when initialization is deferred and needs to be done in a context different from when the object was created. We just want to maintain that object and its property data gave the same context. Otherwise we have inconsistencies in what comes from cached direct accesses and actual property accesses (causes asserts, at least, perhaps worse).