Add SlotArrayCheck in Lowerer instead of IRBuilder
Fixes OS#19767482

We insert LdSlots at the top of the function in jitloopbody for all syms
that are coming in live to the loop. These LdSlots should be restored
correctly on BailOutFromSimpleJitToJitLoopBody.
However we do unreachable code elimination in flowgraph in simplejit,
which can dead code the uses of these syms, and so they will not be
restored on bailout.
This works if we run deadstore pass which will cleanup all the LdSlots
inserted at the top of the function, after we run unreachable block
elimination phase.
But since we turn off deadstore for functions with try/catch, we end up
having a nullptr AV when we do a SlotArrayCheck followed by LdSlot