OS#19405757: Don't reset hasBailedOut bit when we bailout from a non exception finally region
We use hasBailedOut bit to indicate wheter we have to catch the exception or pass the exception up the call stack.
For this, we maintain a stack of hasBailedOut bit ptrs based on exception handling depth via ctors/dtors when we enter/leave an exception handling region.
But, we execute non-exception finally after the dtor pops off the current exception handling region's hasBailedOut bit. Therefore, setting it on bailout would be incorrect, because the top of the stack of hasBailedOut bits is not of the current try-finally, but of its parent.