Fixing Assert at InitBoxedInlineSegments
The assert will be fired as the size of the segment can be larger than length. Say when you add a item on an Array on that size position.
However the assert needs to go away along with that we need to fix the size of new Array with max of INLINE_CHUNK_SIZE, so that IsInlineSegment logic
works correctly as the new Array's segment was indeed inlined.
Truncating the size would not be a problem as this cannot be smaller than length and later if the size needed to be increased the current segment will be
re-allocated.