OS#19078643: Fix constant branch folding when equal sources can be Undefined
We constant fold branches with same sources. The outcome of this folding is different when the source has a valueType of Undefined.
Previously, we checked for this case with ValueType::IsUndefined. But this check does not include valueTypes like - Undefined | Boolean, in which case we cannot constant fold.
With this fix, we handle the case where we can have a definite valueType which could be Undefined.