OS#19035235:Fix constant branch folding for BrOnObject
BrOnObject is true for non-primitives, false for primitives. We were relying on IsDefinite along with IsPrimitive check to constant fold BrOnObject.
But this is not sufficient for a valueType like Null | Object. IsPrimitive returns false because it sees the Object bit, we cannot constant fold BrOnObject to true with just this check.
Because the valueType can also be Null. This change fixes this.