OS#20500679: Disable addition of masking blocks for speculation hoisting when globopt is off
Addition of masking on blocks for speculation hoisting on loop out edges was turned off for SimpleJit due to reduced perf concerns and issues with assigning regions.
However, it was not turned off when globopt is turned off. So we ended up with asserts regarding regions in FlowGraph::Destroy.
This changes fixes this.