Cleanup some region propagation code from flowgraph
We don't really need UpdateRegionFromEHPred. UpdateRegionFromPred is sufficient.
I had added UpdateRegionFromEHPred when we would get different region for a block, based on the predecessor we chose.
But since then, we add all possible flow edges for try/finallys and hence we don't have this issue anymore.
This change does the cleanup.