Implement Numeric Separator
Add support for numeric separator characters in numeric literals. These are just syntactic sugar which improve readability of numeric constants in source. They desugar completely out of the numbers when parsed.

Numeric digits in the constant may have a single '_' character between them. The constant may not begin or end with a '_' character and there may not be multiple numeric separators in a row. All numeric constants are supported. Includes decimal, hex, octal and binary numeric constants.

```javascript
1234 === 1_2_3_4; // true
12.34e56 === 1_2.3_4e5_6; // true
0xff === 0xf_f; // true
0o17 === 0o1_7; // true
0b11 === 0b1_1; // true
```

Numeric Separator proposal is in stage 3 and implemented by JSC and V8. See proposal:
https://github.com/tc39/proposal-numeric-separator

Fixes: #6060