Implement hashbang syntax
Hashbang is a single-line comment syntax designed to mimic interpreter directive syntax common in command-line scripting.

The proposal has reached stage 3 and the syntax is fully supported in SpiderMonkey and V8. See the proposal:
https://github.com/tc39/proposal-hashbang

This a pretty simple syntax with only a couple of rules. The hashbang acts like a single-line comment except the hashbang sequence (`#!`) must be the first token in the source text.