Implement import.meta
The import.meta proposal has reached stage 3. See details:
https://github.com/tc39/proposal-import-meta

This special name is a meta-property which binds to an object in module code. The import.meta object is unique per-module. Each module's import.meta object is filled-out with properties provided by the host. The object itself is lazily-created and an empty object with a null prototype by default.

When loading import.meta for a module, we ask the host to provide a set of properties for the object and allow the host to do other setup of the import.meta object it might need to do. There are two host API provided for in the proposal: HostGetImportMetaProperties and HostFinalizeImportMeta. I've plumbed these callbacks in so hosts can implement them.