Use non-temps for for-of loop control variables
Currently, temps are used to control whether the "return" method on
iterators should be called when exiting from a for-of loop. When for-of
loop bodies are JIT'd, the optimizer only sees the assignment to these
registers and eliminates them as dead stores.

Instead, we'll use non-temps so that the values of these control
variables are written back to the interpreter on exit from the loop
body.