Correctly handle named function expressions when saving ScopeInfo
Previous fix for a similar issue #6157 didn't handle the case of a function expression in which the param and function expression scopes are merged with the body scope. In that case, the function defined in the param scope needs to have access to the body scope of the enclosing function.

Fixes:
https://microsoft.visualstudio.com/OS/_workitems/edit/18926499/