Proxy [[Construct]] trap confuses super call flag
When we call the proxy [[Construct]] method from a JIT helper, it passes an argument in the 0th slot. That makes `JavascriptOperators::GetAndAssertIsConstructorSuperCall` think that we have a `new.target` and are in a super call chain regardless of if we are or are not in a super call chain.

This flag is then used to control constructing the 'this' binding. When we get it wrong, 'this' is bound to the proxy itself instead of a new object.

Fixes:
https://microsoft.visualstudio.com/OS/_workitems/edit/21193960/