Set hasBailedOutBitPtr correctly for nested finally blocks
JITed functions use a stack of bits to determine whether a bailout
occured within a try-catch or try-finally block. When a bailout occurs
within a finally, the corresponding entry in the bit stack has already
been popped, but we still need to set the bit correctly for the
containing try block, if one exists.