Implement nullish coalescing operator
This operator is similar to the logical or operator (`||`) but is meant to avoid unintended behavior which might arise when using logical or. Problem with logical or is that the left argument is coerced to a boolean which means that if it is something falsy (like `0`, `''`, or `false`) then the right argument will be used instead.

The coalescing operator (`??`) is designed to solve this problem. It only assigns the right value when the left value is `null` or `undefined`.

```javascript
null || 'something'; // 'something'
undefined || 'something'; // 'something'
'' || 'something'; // 'something'
0 || 'something'; // 'something'
false || 'something'; // 'something'

null ?? 'something'; // 'something'
undefined ?? 'something'; // 'something'
'' ?? 'something'; // ''
0 ?? 'something'; // 0
false ?? 'something'; // false
```