Correctly set flag indicating a function has references to `this` symbol
`IsArgSaveRequired` depends on the flag that the function has references to `this` (amongst other conditions) to determine if we need to spill arguments to stack on x64. Since this flag is not set correctly, sometimes we end up not spilling arguments, and therefore have incorrect argument values. Also remove previously unreachable code due to the flag not being set.