Only emit dummy yield when a function has non-simple parameter list
Generator functions with non-simple parameter list (e.g: arguments with default values that have some complex expressions) will need to evaluate those expressions when the generator object is created. This result in us emitting a first dummy yield to evaluate those. For jitted generator functions, this would mean every function would have a bailout at the beginning which is fairly expensive for small functions. So to mitigate this, only emit dummy yields when we really need to do so.