Fix for the array inline segment assert.
This assert is happening due to false positive. The array's head segment is not allocated as inlined - but it was allocated on the page boundary.
That makes this segement to just aligned with the array and we got the assert fired for wrong reason.
We don't need to check the assert for the source array to be not inlined as in the above 'else' condition we allocated the dest segment ourselve in the same function.
So I removed the part where we check the current array is not inlined.