Issue in setting accessor in PathTypeHandlerWithAttr
When the property is deleted (say when we are in SimpleTypeHandler) and then later
if we create an accessor on the same property we have converted the type handler to PathTypeHandlerWithAttr.
Since a property got deleted and PathTypeHandler should not handle that deleted property we need to convert this to
Dictionary type.
In order to fix that When we add the accessor on any property we need to check if any property deleted upon that we need
to convert to Dictionary type instead of type.