Enable Bytecode regen on Linux/MacOS (#6425)
Introduce a Force32BitByteCode flag usable in Debug and Test builds that forces a 64bit CC build to generate 32bit bytecode for the known differences. (All related logic is #ifdef'd out in release builds)

Add a python script that will use that flag to perform the regen on macos & linux.