Correct ifdef for arm64 (#6995)
JsDisableRuntimeExecution currently does not function on AppleSilicon.

This is because of an incorrect ifdef, JsDisableRuntimeExecution works by overriding CC's stack size tracking so that CC mistakenly thinks there is no stack space left and hence stops executing JS; in order to do this StackLimitForScriptInterrupt needs to be an appropriately large number for the architecture, currently it's being set to the same size on AppleSilicon that was used for x86 which is obviously not high enough, change the ifdef so all 64bit architectures (i.e. x64 and AppleSilicon) get the larger figure.

Bug reported by @fatcerberus fix also tested offline by @fatcerberus

Currently most of our native API tests only run on windows hence do not detect this; we should look at enabling them on macOS/Linux as a future todo.