Merging PR: https://github.com/chakra-core/ChakraCore/pull/6531.
Merging PR: https://github.com/chakra-core/ChakraCore/pull/6531. 
This is the change that was added by Microsoft as part of their December 2020 Security Update, which addresses the CVE-2020-17131. This CVE is to address Out-of-bounds Write in ChakraCore, and there by fixing memory corruption vulnerability. This change is unfortunately missed in the latest ChakraCore code base, and hence adding this for review.